pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7.0"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.25.1"
selenium = "^3.141.0"
pytest = "^6.2.4"
gunicorn = "^20.1.0"
pymongo = {extras = ["srv"], version = "^4.2.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"

